move transformation from csv to zeitintervall from ZaehlungDialog to KnotenLageForm
